fixed `emulate`
