+++ scripts/__init__.pyi
@@ -148,7 +148,9 @@ class Section:
     def is_readable(self) -> bool: ...
     def is_writable(self) -> bool: ...
     def is_executable(self) -> bool: ...
+    @property
     def size(self) -> int: ...
+    @property
     def realpath(self) -> str: ...
     def __str__(self) -> str: ...
 
@@ -1645,6 +1647,17 @@ class GefSettingsManager(dict):
 
 
 class GefSessionManager(GefManager):
+    remote: Optional[bool] = ...
+    qemu_mode: bool = ...
+    convenience_vars_index: int = ...
+    heap_allocated_chunks: List[Tuple[int, int]] = ...
+    heap_freed_chunks: List[Tuple[int, int]] = ...
+    heap_uaf_watchpoints: List[UafWatchpoint] = ...
+    pie_breakpoints: Dict[int, PieVirtualBreakpoint] = ...
+    pie_counter: int = ...
+    aliases: List[GefAlias] = ...
+    modules: List[Elf] = ...
+    constants: Dict[str, pathlib.Path] = ...
     def __init__(self) -> None: ...
     def reset_caches(self) -> None: ...
     @property
